Final commit

+++ Polygons/Assets/Scripts/Shaders/BGEffect.shader
@@ -56,7 +56,7 @@
 
                 float isStripe = 1000 * (y - 0.5);
                 isStripe = saturate(isStripe);
-                float4 newCol = lerp(col, _StripeCol, _NonStripeAlpha);
+                fixed4 newCol = lerp(col, _StripeCol, _NonStripeAlpha);
                 col = lerp(_StripeCol, newCol, isStripe);
 
                 return col;
